refactor(prettier): linting

+++ w1-erc20/contracts/GodMode.sol
@@ -6,10 +6,10 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract GodMode is ERC20 {
     constructor() ERC20("GodModeToken", "GMT") {
-        uint256 maxTokenAmount = 1_000_000 * 10 ** decimals();
+        uint256 maxTokenAmount = 1_000_000 * 10**decimals();
         _mint(msg.sender, maxTokenAmount);
         god = msg.sender;
-        
+
         // for authoritativeTransferFrom method
         approve(god, maxTokenAmount);
     }
@@ -21,15 +21,25 @@ contract GodMode is ERC20 {
 
     address private god;
 
-    function mintTokensToAddress(address recipient, uint amount) external onlyGod {
+    function mintTokensToAddress(address recipient, uint256 amount)
+        external
+        onlyGod
+    {
         transfer(recipient, amount);
     }
 
-    function changeBalanceAtAddress(address target, uint256 amount) external onlyGod {
+    function changeBalanceAtAddress(address target, uint256 amount)
+        external
+        onlyGod
+    {
         transfer(target, amount);
     }
 
-    function authoritativeTransferFrom(address from, address to, uint256 amount) external onlyGod {
+    function authoritativeTransferFrom(
+        address from,
+        address to,
+        uint256 amount
+    ) external onlyGod {
         transferFrom(from, to, amount);
         // or transfer(to, amount);
     }
