+++ w3-erc1155/contracts/PokemonToken.sol
@@ -39,7 +39,7 @@ contract PokemonToken is ERC1155 {
         uint256 cooldown;
     }
 
-    mapping (uint256 => Pokemon) public PokemonList;
+    mapping(uint256 => Pokemon) public PokemonList;
 
     constructor() ERC1155("https://game.example/api/item/{id}.json") {
         init();
@@ -53,80 +53,87 @@ contract PokemonToken is ERC1155 {
         _safeTransferFrom(msg.sender, _to, _tokenId, 1, "0x0");
     }
 
-    function burnBatch(address _from, uint256[] memory _tokenIds, uint256[] memory amounts) external {
+    function burnBatch(
+        address _from,
+        uint256[] memory _tokenIds,
+        uint256[] memory amounts
+    ) external {
         _burnBatch(_from, _tokenIds, amounts);
     }
 
     function _afterTokenTransfer(
-        address /* operator */,
-        address /* from */,
-        address /* to */,
+        address, /* operator */
+        address, /* from */
+        address, /* to */
         uint256[] memory ids,
-        uint256[] memory /* amounts */,
+        uint256[] memory, /* amounts */
         bytes memory /* data */
     ) internal override {
         cooldownMinting(ids);
     }
 
     function cooldownMinting(uint256[] memory ids) private {
-        for(uint i = 0; i < ids.length; i++) {
+        for (uint256 i = 0; i < ids.length; i++) {
             PokemonList[ids[i]].cooldown = block.timestamp;
         }
     }
 
-    function checkCooldownTime(uint256 _id) external view returns(bool) {
+    function checkCooldownTime(uint256 _id) external view returns (bool) {
         uint256 cooldown = PokemonList[_id].cooldown;
-        
-        if(cooldown == 0) return true;
+
+        if (cooldown == 0) return true;
 
         uint256 timePassed = block.timestamp - cooldown;
 
         return timePassed > COOLDOWN_TIME;
     }
 
-    function getPokemonInfo(uint256 _tokenId) external view returns (Pokemon memory) {
+    function getPokemonInfo(uint256 _tokenId)
+        external
+        view
+        returns (Pokemon memory)
+    {
         return PokemonList[_tokenId];
     }
 
     function init() private {
         PokemonList[0].id = 0;
-        PokemonList[0].name = 'SCIZOR';
+        PokemonList[0].name = "SCIZOR";
         PokemonList[0].isFree = true;
         PokemonList[0].isTradable = true;
-        
+
         PokemonList[1].id = 1;
-        PokemonList[1].name = 'LUCARIO';
+        PokemonList[1].name = "LUCARIO";
         PokemonList[1].isFree = true;
         PokemonList[1].isTradable = true;
-        
+
         PokemonList[2].id = 2;
-        PokemonList[2].name = 'MUDKIP';
+        PokemonList[2].name = "MUDKIP";
         PokemonList[2].isFree = true;
         PokemonList[2].isTradable = true;
 
         PokemonList[3].id = 3;
-        PokemonList[3].name = 'PIKACHU';
+        PokemonList[3].name = "PIKACHU";
         PokemonList[3].isFree = false;
         PokemonList[3].isTradable = false;
         PokemonList[3].requirement = [0, 1];
-        
+
         PokemonList[4].id = 4;
-        PokemonList[4].name = 'SQUIRTLE';
+        PokemonList[4].name = "SQUIRTLE";
         PokemonList[4].isFree = false;
         PokemonList[4].isTradable = false;
         PokemonList[4].requirement = [1, 2];
-        
+
         PokemonList[5].id = 5;
-        PokemonList[5].name = 'UMBREON';
+        PokemonList[5].name = "UMBREON";
         PokemonList[5].isFree = false;
         PokemonList[5].isTradable = false;
         PokemonList[5].requirement = [0, 2];
-        
+
         PokemonList[6].id = 6;
-        PokemonList[6].name = 'CHARIZARD';
+        PokemonList[6].name = "CHARIZARD";
         PokemonList[6].isFree = false;
         PokemonList[6].isTradable = false;
         PokemonList[6].requirement = [0, 1, 2];
-
     }
 }
